+++ lib/verquest/properties/field.rb
@@ -6,6 +6,7 @@ module Verquest
     #
     # Represents simple scalar types (string, number, integer, boolean) in the schema.
     # Used for defining basic data fields without nesting.
+    # Supports both default types and custom field types defined in the configuration.
     #
     # @example Define a required string field
     #   field = Verquest::Properties::Field.new(
@@ -15,28 +16,39 @@ module Verquest
     #     format: "email"
     #   )
     class Field < Base
-      # List of allowed field types
+      # List of default field types
       # @return [Array<Symbol>]
-      ALLOWED_TYPES = %w[string number integer boolean].freeze
+      DEFAULT_TYPES = %w[string number integer boolean].freeze
 
       # Initialize a new Field property
       #
       # @param name [String, Symbol] The name of the property
-      # @param type [String, Symbol] The data type for this field, must be one of ALLOWED_TYPES
-      # @param required [Boolean] Whether this property is required
+      # @param type [String, Symbol] The data type for this field, can be a default type or a custom field type
+      # @param required [Boolean] Whether this property is required (overridden by custom type if it defines required)
       # @param map [String, nil] The mapping path for this property
-      # @param schema_options [Hash] Additional JSON schema options for this property
-      # @raise [ArgumentError] If type is not one of the allowed types
+      # @param schema_options [Hash] Additional JSON schema options for this property (merged with custom type options)
+      # @raise [ArgumentError] If type is not one of the allowed types (default or custom)
       # @raise [ArgumentError] If attempting to map a field to root without a name
       def initialize(name:, type:, required: false, map: nil, **schema_options)
-        raise ArgumentError, "Type must be one of #{ALLOWED_TYPES.join(", ")}" unless ALLOWED_TYPES.include?(type.to_s)
+        raise ArgumentError, "Type must be one of #{allowed_types.join(", ")}" unless allowed_types.include?(type.to_s)
         raise ArgumentError, "You can not map fields to the root without a name" if map == "/"
 
+        if (custom_type = Verquest.configuration.custom_field_types[type.to_sym])
+          @type = custom_type[:type].to_s
+          @required = custom_type.key?(:required) ? custom_type[:required] : required
+          @schema_options = if custom_type.key?(:schema_options)
+            custom_type[:schema_options].merge(schema_options).transform_keys(&:to_s)
+          else
+            schema_options.transform_keys(&:to_s)
+          end
+        else
+          @type = type.to_s
+          @required = required
+          @schema_options = schema_options&.transform_keys(&:to_s)
+        end
+
         @name = name.to_s
-        @type = type.to_s
-        @required = required
         @map = map
-        @schema_options = schema_options&.transform_keys(&:to_s)
       end
 
       # Generate JSON schema definition for this field
@@ -60,6 +72,13 @@ module Verquest
       private
 
       attr_reader :type, :schema_options
+
+      # Gets the list of allowed field types, including both default and custom types
+      #
+      # @return [Array<String>] Array of allowed field type names
+      def allowed_types
+        DEFAULT_TYPES + Verquest.configuration.custom_field_types.keys.map(&:to_s)
+      end
     end
   end
 end