+++ test/verquest/properties/field_test.rb
@@ -83,6 +83,35 @@ module Verquest
 
         assert_equal expected_mapping, mapping
       end
+
+      def test_custom_field_type
+        Verquest.configuration.custom_field_types = {
+          custom_type: {
+            type: "string",
+            schema_options: {format: "custom", pattern: /\Acustom_\w+\z/, min_length: 5, max_length: 20}
+          }
+        }
+
+        field = Field.new(
+          name: :custom_field,
+          type: :custom_type,
+          required: true,
+          description: "A custom field"
+        )
+
+        expected_schema = {
+          "custom_field" => {
+            "type" => "string",
+            "format" => "custom",
+            "pattern" => /\Acustom_\w+\z/,
+            "minLength" => 5,
+            "maxLength" => 20,
+            "description" => "A custom field"
+          }
+        }
+
+        assert_equal expected_schema, field.to_schema
+      end
     end
   end
 end